Restored
This is the restored files of the map I was working on before I broke the last fork.

Changes:

Fixed improper names of the map name to the proper one.

The station now has a name! NTB Prometheus the brother of Atlas

Made a separate planet weather setup for the location of this map since Prometheus is supposed to be in a more severe climate compared to Atlas